2005-01-18	Ralf Corsepius <ralf.corsepius@rtems.org>
	* sapi/include/rtems/io.h : size_t device_name_length.

+++ cpukit/sapi/include/rtems/io.h
@@ -74,7 +74,7 @@ typedef struct {
 
 typedef struct {
     char                     *device_name;
-    uint32_t                  device_name_length;
+    size_t                    device_name_length;
     rtems_device_major_number major;
     rtems_device_minor_number minor;
 } rtems_driver_name_t;
